+++ src/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2022-04-05
+
+* [RELEASE STABLE] Released as Stable
+* [RELEASE DEV] Released as Dev
+
+* [Fix] Fix the invoice discount entered as percentage is set vat inclusive even if the invoice is vat exclusive.
+
 ## 2022-04-03
 
 * [RELEASE STABLE] Released as Stable